ml_hair/requirements.txt
flask
pillow
numpy
torch
